Add files via upload
#PS script upload. Author : Mayank Tripathi

Developed highly automated PowerShell script binded with batch script, which was setup with Scheduler to trigger my PS script. The PS script collect all the server Health report in csv for multiple domain. 

A csv file with detail report was sent along with Replication result. The script also produce the outfile for Trust (one way & two way) for inter forest AD environment. 

The Health Check result was further used to published the data on internal dashboard for the presentation purpose for the server uptime and performance monitoring with Engineering Team.